refactor: Add error handling to CarImpService for database exceptions

+++ src/main/java/br/com/lux/services/car/carImp/CarImpService.java
@@ -5,11 +5,13 @@ import br.com.lux.domain.car.CarPageType;
 import br.com.lux.repository.car.CarRepository;
 import br.com.lux.services.car.CarService;
 
+import br.com.lux.services.exception.ServiceException;
 import jakarta.validation.ConstraintViolation;
 import jakarta.validation.ConstraintViolationException;
 import jakarta.validation.Validator;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
@@ -38,49 +40,99 @@ public class CarImpService implements CarService
     @Transactional(propagation = Propagation.NOT_SUPPORTED)
     public List<Car> findCarAll()
     {
-        return carRepository.findAll();
+        try
+        {
+            return carRepository.findAll();
+        }
+        catch (Exception e)
+        {
+            throw new ServiceException("Erro ao buscar todos os carros! " + e.getMessage());
+        }
     }
 
     @Override
+    @Transactional
     public void registerCar(Car car)
     {
-        Set<ConstraintViolation<Car>> violations = validator.validate(car);
-        if (!violations.isEmpty()) {
-            throw new ConstraintViolationException(violations);
-        }
+        try
+        {
+            Set<ConstraintViolation<Car>> violations = validator.validate(car);
+            if (!violations.isEmpty()) {
+                throw new ConstraintViolationException(violations);
+            }
 
-        carRepository.save(car);
+            carRepository.save(car);
+        }
+        catch (DataIntegrityViolationException e)
+        {
+            throw new ServiceException("Erro ao cadastrar o carro! " + e.getMessage());
+        }
     }
 
     @Override
     @Transactional(propagation = Propagation.NOT_SUPPORTED)
     public Car findCarById(Integer id)
     {
-        return carRepository.findById(id).orElse(null);
+        try
+        {
+            return carRepository.findById(id).orElse(null);
+        }
+        catch (Exception e)
+        {
+            throw new ServiceException("Erro ao buscar o carro! " + e.getMessage());
+        }
     }
 
     @Override
+    @Transactional
     public void deleteCar(Integer id)
     {
-        carRepository.deleteById(id);
+        try
+        {
+            if (id == null)
+                throw new ServiceException("Id inválido!");
+
+            carRepository.deleteById(id);
+        }
+        catch (ServiceException e)
+        {
+            throw new ServiceException("Erro ao deletar o carro por id! " + e.getMessage());
+        }
     }
 
     @Override
     @Transactional(propagation = Propagation.NOT_SUPPORTED)
     public long countCars()
     {
-        return carRepository.count();
+        try
+        {
+            return carRepository.count();
+        }
+        catch (Exception e)
+        {
+            throw new ServiceException("Erro ao contar os carros! " + e.getMessage());
+        }
     }
 
+    @Override
+    @Transactional(propagation = Propagation.NOT_SUPPORTED)
     public Map<CarPageType, Long> getCarTypeCounts()
     {
-        Map<CarPageType, Long> carTypeCounts = new HashMap<>();
-
-        for (CarPageType type : CarPageType.values())
+        try
         {
-            carTypeCounts.put(type, carRepository.countByPage(type));
+            Map<CarPageType, Long> carTypeCounts = new HashMap<>();
+
+            for (CarPageType type : CarPageType.values())
+            {
+                carTypeCounts.put(type, carRepository.countByPage(type));
+            }
+
+            return carTypeCounts;
+        }
+        catch (Exception e)
+        {
+            throw new ServiceException("Erro ao contar os tipos de carros! " + e.getMessage());
         }
 
-        return carTypeCounts;
     }
 }
